Gems/Atom

+++ Gems/Atom/Feature/Common/Code/Source/ReflectionScreenSpace/ReflectionScreenSpaceBlurChildPass.cpp
@@ -44,8 +44,8 @@ namespace AZ
             }
 
             float inverseScale = 1.0f / m_outputScale;
-            uint32_t outputWidth = m_imageSize.m_width * inverseScale;
-            uint32_t outputHeight = m_imageSize.m_height * inverseScale;
+            uint32_t outputWidth = static_cast<uint32_t>(m_imageSize.m_width * inverseScale);
+            uint32_t outputHeight = static_cast<uint32_t>(m_imageSize.m_height * inverseScale);
 
             params.m_viewportState = RHI::Viewport(0, static_cast<float>(outputWidth), 0, static_cast<float>(outputHeight));
             params.m_scissorState = RHI::Scissor(0, 0, outputWidth, outputHeight);
